+++ RaiseMeUp/RaiseMeUp/Sources/Core/Constant/KeychainError.swift
@@ -1,0 +1,15 @@
+//
+//  KeychainError.swift
+//  RaiseMeUp
+//
+//  Created by 홍석현 on 11/27/23.
+//
+
+import Foundation
+
+enum KeychainError: Error {
+    case noPassword
+    case unexpectedPasswordData
+    case unexpectedItemData
+    case unhandledError
+}
